pyproject.toml
[tool.poetry]
name = "power of 3"
version = "0.1.0"
description = "Forex Trader"
authors = ["John Thadison"]

[tool.poetry.dependencies]
python = ">=3.10.0,<=3.12.10"
crewai = "0.121.0"
unstructured = '==0.10.25'
pyowm = '3.3.0'
python-dotenv = "1.0.0"
langchain-openai = "0.3.18"
yfinance = "^0.2.40"
sqlalchemy = "^2.0.41"
pandas-ta = "^0.3.14b0"
langchain = "^0.3.25"
psycopg2 = "^2.9.10"
oandapyv20 = "^0.7.2"
schedule = "^1.2.2"
requests = "^2.32.3"
psycopg2-binary = "^2.9.10"
pandas = "^2.2.3"
numpy = "^2.2.6"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"